refactor: 최종 승패 출력 기능 이전까지의 모든 기능들을 v1 패키지로 이관한다
