+++ src/main/kotlin/game/blackjack/v1/domain/Blackjack.kt
@@ -1,4 +1,4 @@
-package game.blackjack.domain
+package game.blackjack.v1.domain
 
 object Blackjack {
     const val WINNING_SCORE: Int = 21
