+++ src/main/kotlin/game/blackjack/v1/domain/Card.kt
@@ -1,4 +1,4 @@
-package game.blackjack.domain
+package game.blackjack.v1.domain
 
 data class Card(val number: CardNumber, val shape: CardShape) {
     override fun toString() = "${number.value}${shape.value}"
