+++ src/test/kotlin/game/blackjack/v1/domain/CardNumberTest.kt
@@ -1,18 +1,18 @@
-package game.blackjack.domain
+package game.blackjack.v1.domain
 
-import game.blackjack.domain.CardNumber.ACE
-import game.blackjack.domain.CardNumber.EIGHT
-import game.blackjack.domain.CardNumber.FIVE
-import game.blackjack.domain.CardNumber.FOUR
-import game.blackjack.domain.CardNumber.JACK
-import game.blackjack.domain.CardNumber.KING
-import game.blackjack.domain.CardNumber.NINE
-import game.blackjack.domain.CardNumber.QUEEN
-import game.blackjack.domain.CardNumber.SEVEN
-import game.blackjack.domain.CardNumber.SIX
-import game.blackjack.domain.CardNumber.TEN
-import game.blackjack.domain.CardNumber.THREE
-import game.blackjack.domain.CardNumber.TWO
+import game.blackjack.v1.domain.CardNumber.ACE
+import game.blackjack.v1.domain.CardNumber.EIGHT
+import game.blackjack.v1.domain.CardNumber.FIVE
+import game.blackjack.v1.domain.CardNumber.FOUR
+import game.blackjack.v1.domain.CardNumber.JACK
+import game.blackjack.v1.domain.CardNumber.KING
+import game.blackjack.v1.domain.CardNumber.NINE
+import game.blackjack.v1.domain.CardNumber.QUEEN
+import game.blackjack.v1.domain.CardNumber.SEVEN
+import game.blackjack.v1.domain.CardNumber.SIX
+import game.blackjack.v1.domain.CardNumber.TEN
+import game.blackjack.v1.domain.CardNumber.THREE
+import game.blackjack.v1.domain.CardNumber.TWO
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.data.forAll
 import io.kotest.data.row
